move lcd_sdk to main package
