+++ tests/test_lcd_sdk.py
@@ -13,7 +13,7 @@ from pytest import mark
                                              ('logi_lcd_color_set_title', ('', (1, 2, 3)), False),
                                              ('logi_lcd_color_set_text', (1, '', (1, 2, 3)), False)])
 def test_all_failure_cases(function, args, result):
-    from dcspy.sdk import lcd_sdk
+    from dcspy import lcd_sdk
     lcd_sdk.LCD_DLL = None
     assert getattr(lcd_sdk, function)(*args) is result
 
@@ -29,7 +29,7 @@ def test_all_failure_cases(function, args, result):
                                                     ('logi_lcd_color_set_title', 'LogiLcdColorSetTitle', ('', (1, 2, 3)), True),
                                                     ('logi_lcd_color_set_text', 'LogiLcdColorSetText', (1, '', (1, 2, 3)), True)])
 def test_all_success_cases(py_func, c_func, args, result):
-    from dcspy.sdk import lcd_sdk
+    from dcspy import lcd_sdk
     mocked_c_func = Mock()
     mocked_c_func.return_value = result
     lcd_sdk.LCD_DLL = {c_func: mocked_c_func}
@@ -44,7 +44,7 @@ def test_all_success_cases(py_func, c_func, args, result):
                                              (2, 320 * 240, [call(0, ''), call(1, ''), call(2, ''), call(3, ''),
                                                              call(4, ''), call(5, ''), call(6, ''), call(7, '')]))])
 def test_clear_display(c_funcs, effect, args):
-    from dcspy.sdk import lcd_sdk
+    from dcspy import lcd_sdk
     with patch.object(lcd_sdk, 'logi_lcd_is_connected', side_effect=effect) as connected:
         with patch.object(lcd_sdk, c_funcs[0], return_value=True) as set_background:
             with patch.object(lcd_sdk, c_funcs[1], return_value=True) as set_text:
