+++ tests/test_logitech.py
@@ -5,7 +5,7 @@ def test_check_keyboard_mono_has_aircraft_member():
     from dcspy.logitech import KeyboardMono
     from dcspy.dcsbios import ProtocolParser
     from dcspy.aircrafts import Aircraft
-    from dcspy.sdk import lcd_sdk
+    from dcspy import lcd_sdk
 
     with patch.object(lcd_sdk, 'logi_lcd_init', return_value=True):
         lcd = KeyboardMono(ProtocolParser())
@@ -16,7 +16,7 @@ def test_check_keyboard_color_has_aircraft_member():
     from dcspy.logitech import KeyboardColor
     from dcspy.dcsbios import ProtocolParser
     from dcspy.aircrafts import Aircraft
-    from dcspy.sdk import lcd_sdk
+    from dcspy import lcd_sdk
 
     with patch.object(lcd_sdk, 'logi_lcd_init', return_value=True):
         lcd = KeyboardColor(ProtocolParser())
